requirements.txt
absl-py==0.11.0
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1606255208697/work
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
cycler==0.10.0
dataclasses==0.6
Flask @ file:///home/ktietz/src/ci/flask_1611932660458/work
gevent @ file:///C:/ci/gevent_1610977717367/work
greenlet @ file:///C:/ci/greenlet_1611958565931/work
imutil==0.2.5
itsdangerous @ file:///home/ktietz/src/ci/itsdangerous_1611932585308/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mediapipe==0.8.2
numpy==1.19.3
opencv-python==4.5.1.48
Pillow==8.0.1
protobuf==3.14.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
wincertstore==0.2
ws4py==0.5.1
zope.event==4.5.0
zope.interface @ file:///C:/ci/zope.interface_1606940376312/work
